update - works fine, but could be cleaner

+++ src/GoFetch.java
@@ -21,16 +21,26 @@ import java.util.Map;
 public class GoFetch {
 
 	/**
+	 * Wrong command line error message.
+	 */
+	private static final String WRONG_COMMAND_LINE = "Wrong command line. Please specify a correct command line : ";
+	/**
+	 * Example of command line.
+	 */
+	private static final String ARG_EXAMPLE = "EXAMPLE : (java) GoFetch --url http://www.website.com/img[X].jpg --target C:/folder/a/ --range 1-6";
+
+	/**
 	 * Enum holding all the possible arguments for this program.
 	 * 
 	 * @author lmadeuf
-	 * 
 	 */
 	private enum Argument {
 
-		URL("-u", "--url", "TODO help of url"),
-		TARGET("-t", "--target", "TODO help of target"),
-		RANGE("-r", "--range", "TODO help of range");
+		URL("-u", "--url", "-u/--url		url argument, the url of the doxuments to download. Replace one integer occurence with '[X]', wich will be replaced by the numbers from the specified range."),
+		TARGET("-t", "--target", "-t/--target		target argument, the value is a folder path on disk, required to write the files."),
+		RANGE("-r", "--range", "-r/--range		range argument, requires a specific syntax value 'n1:n2' where n1 and n2 are two positive integers >= 0 and n1 >= n2."),
+		HELP("-h", "--help", "-h/--help		help argument without value, will display the help of the program and shut it down."),
+		SIMULATION("-s", "--simul", "-s/--simul		simul argument, no value required. If present, the files will not really be downloaded or written to disk, it will fake everything for debug / demonstration purposes.");
 
 		/**
 		 * The short version of the argument name.
@@ -50,9 +60,15 @@ public class GoFetch {
 		private Argument(final String shortLabel, final String longLabel, final String helpLine) {
 			this.shortLabel = shortLabel;
 			this.longLabel = longLabel;
+			this.helpLine = helpLine;
 		}
 
-		public static Argument resolve(final String argline) {
+		/**
+		 * Recolves a correct Argument enum value from it's string representation.
+		 * @param argline the argument from the command line to be recognised.
+		 * @return the Argument enum value found, if a match was detected.
+		 */
+		private static Argument resolve(final String argline) {
 			Argument result = null;
 			for (final Argument arg : Argument.values()) {
 				if (arg.shortLabel.equals(argline) || arg.longLabel.equals(argline)) {
@@ -63,36 +79,57 @@ public class GoFetch {
 			return result;
 		}
 		
-		public String toString(){
+		/**
+		 * Retrieves the help text from the argument enum value.
+		 */
+		public String getHelpLine(){
 			return helpLine;
 		}
 	}
 
+	/**
+	 * Where the magic happens.
+	 * @param args the command line from the user.
+	 */
 	public static void main(final String[] args) {
 		final Map<Argument, String> arguments = resolveCommandLine(args);
-		if (arguments != null) {
-			final List<String> urlList = buildUrlList(arguments);
-			if (urlList != null) {
-				for(final String url : urlList){
-					try {
-						final byte[] data = download(new URL(url));
-						writeToDisk(data, buildTargetFileName(arguments.get(Argument.TARGET), url));
-					} catch (final IOException e) {
-						e.printStackTrace();
-						break;
+		if (isCommandLineOK(arguments)) {
+			if(!arguments.containsKey(Argument.HELP)){
+				final List<String> urlList = buildUrlList(arguments);
+				if (urlList != null) {
+					for(final String url : urlList){
+						try {
+							boolean simulation = arguments.containsKey(Argument.SIMULATION);
+							final byte[] data = download(new URL(url), simulation);
+							writeToDisk(data, buildTargetFileName(arguments.get(Argument.TARGET), url), simulation);
+						} catch (final IOException e) {
+							e.printStackTrace();
+							break;
+						}
 					}
 				}
+			} else {
+				printHelp();
 			}
+		} else {
+			System.out.println(WRONG_COMMAND_LINE);
+			printHelp();
 		}
 	}
 
-	private static String buildTargetFileName(String target, final String url) {
+	/**
+	 * Builds the full filepath of the target file by getting the filename from the compiled
+	 * url and the filepath from the target path.
+	 * @param target target file path
+	 * @param url the dowload url
+	 * @return the String representation of the target filename
+	 */
+	private static String buildTargetFileName(final String target, final String url) {
 		final StringBuilder builder = new StringBuilder();
 		builder.append(target);
 		if(!target.endsWith("/") && target.contains("/")){
 			builder.append("/");
-		}
-		if(!target.endsWith("\\") && target.contains("\\")){
+		} else if(!target.endsWith("\\") && target.contains("\\")){
 			builder.append("\\");
 		}
 		final String[] split = url.split("/");
@@ -120,26 +157,28 @@ public class GoFetch {
 	/**
 	 * Resolves the command line and builds a map with all the arguments found.
 	 * 
-	 * @param args
-	 *            the command line arguments.
+	 * @param args the command line arguments.
 	 * @return the map built.
 	 */
 	private static Map<Argument, String> resolveCommandLine(final String[] args) {
 		Map<Argument, String> result = null;
-		if (args == null || args.length == 0 || args.length != 6) {
-			System.out.println("No arguments specified. Please specify all of these : ");
+		if (args == null || args.length == 0) {
+			System.out.println(WRONG_COMMAND_LINE);
 			printHelp();
 		} else {
 			result = new HashMap<Argument, String>();
 			final Iterator<String> it = Arrays.asList(args).iterator();
-	
 			while (it.hasNext()) {
-				String arg = (String) it.next();
+				final String arg = (String) it.next();
 				final Argument argument = Argument.resolve(arg);
 				if (argument != null) {
-					result.put(argument, (String) it.next());
+					if(Argument.HELP.equals(argument) || Argument.SIMULATION.equals(argument)){
+						result.put(argument, "");
+					}else{
+						result.put(argument, (String) it.next());
+					}
 				} else {
-					System.out.println("Wrong command line. Please specify all of these : ");
+					System.out.println(WRONG_COMMAND_LINE);
 					printHelp();
 					result = null;
 					break;
@@ -150,12 +189,31 @@ public class GoFetch {
 	}
 
 	/**
+	 * Checks that the command line is usable as parsed.
+	 * @param commandMap the map containing the arguments and their values.
+	 * @return true if the command line is usable as parsed.
+	 */
+	private static boolean isCommandLineOK(final Map<Argument, String> commandMap) {
+		boolean isCommandLineOK = true;
+		if(commandMap==null || commandMap.isEmpty()){
+			isCommandLineOK = false;
+		} else if(!commandMap.containsKey(Argument.HELP) && (commandMap.size()<3)){
+			isCommandLineOK = false;
+		} else if(!commandMap.containsKey(Argument.HELP) && !commandMap.containsKey(Argument.RANGE) 
+				&& !commandMap.containsKey(Argument.URL) && !commandMap.containsKey(Argument.TARGET)){
+			isCommandLineOK = false;
+		}
+		return isCommandLineOK;
+	}
+
+	/**
 	 * Prints out the help contents.
 	 */
 	private static void printHelp() {
 		for (final Argument arg : Argument.values()) {
-			System.out.println(arg.toString());
+			System.out.println(arg.getHelpLine());
 		}
+		System.out.println(ARG_EXAMPLE);
 	}
 
 	/**
@@ -165,11 +223,12 @@ public class GoFetch {
 	 *            the byte[] containing the data of the downloaded file.
 	 * @param fullPathName
 	 *            the path name on disk, ended by the file name.
+	 * @param simulation 
 	 * @throws IOException
 	 *             If an IOException is thrown by FileOutputStream (opening,
 	 *             writing, closing).
 	 */
-	private static void writeToDisk(final byte[] data, final String fullPathName)
+	private static void writeToDisk(final byte[] data, final String fullPathName, final boolean simulation)
 			throws IOException {
 		if (fullPathName == null || fullPathName.length() == 0) {
 			throw new IllegalArgumentException("fullPathName null or empty");
@@ -178,11 +237,13 @@ public class GoFetch {
 			throw new IllegalArgumentException("data null or empty");
 		}
 		System.out.println(String.format("Writing to : %s", fullPathName));
-		final FileOutputStream out = new FileOutputStream(fullPathName);
-		try {
-			out.write(data);
-		} finally {
-			out.close();
+		if(!simulation){
+			final FileOutputStream out = new FileOutputStream(fullPathName);
+			try {
+				out.write(data);
+			} finally {
+				out.close();
+			}
 		}
 	}
 
@@ -192,33 +253,39 @@ public class GoFetch {
 	 * 
 	 * @param url
 	 *            the url to download a file from.
+	 * @param simulation if true, the file will not really be downloaded
 	 * @return byte[] the raw data
 	 * @throws IOException
 	 *             if less bytes have been received than expected, or if read()
 	 *             from {@link java.io.InputStream} throws it, or if
 	 *             openConnection() from {@link java.net.URL} throws it.
 	 */
-	private static byte[] download(final URL url) throws IOException {
+	private static byte[] download(final URL url, final boolean simulation) throws IOException {
+		byte[] data = null;
 		System.out.println(String.format("Downloading from : %s", url));
-		final URLConnection uc = url.openConnection();
-		final int len = uc.getContentLength();
-		final InputStream is = new BufferedInputStream(uc.getInputStream());
-		try {
-			byte[] data = new byte[len];
-			int offset = 0;
-			while (offset < len) {
-				int read = is.read(data, offset, data.length - offset);
-				if (read < 0) {
-					break;
+		if(!simulation){
+			final URLConnection uc = url.openConnection();
+			final int len = uc.getContentLength();
+			final InputStream is = new BufferedInputStream(uc.getInputStream());
+			try {
+				data = new byte[len];
+				int offset = 0;
+				while (offset < len) {
+					int read = is.read(data, offset, data.length - offset);
+					if (read < 0) {
+						break;
+					}
+					offset += read;
 				}
-				offset += read;
+				if (offset < len) {
+					throw new IOException(String.format("Read %d bytes; expected %d", offset, len));
+				}
+			} finally {
+				is.close();
 			}
-			if (offset < len) {
-				throw new IOException(String.format("Read %d bytes; expected %d", offset, len));
-			}
-			return data;
-		} finally {
-			is.close();
+		} else {
+			data = new byte[1];
 		}
+		return data;
 	}
 }
